change the way standard exit id is computed
Instead of relying utxo_pos alone, compute exit id as a function of tx hash.
This allows to locate existing standard exits when caller is providing
only transaction itself, as in case of in-flight exits.

This also changes the way priority value of exit is computed.

Anatomy of priority value, most significant bits first:

42 bits - timestamp (exitable_at)
54 bits - blknum * 10^9 + txindex; own or youngest input (for in-flights)
8 bits - oindex; zero for in-flight
1 bit - in-flight flag
151 bit - tx hash

Anatomy of exit_id (both in-flight and standard):

8 bits - oindex; zero for in-flight
1 bit - in-flight flag
151 bit - tx hash